Ai_RakamTanıma_Projesi/requirements.txt
pandas==1.3.3
scikit-learn==0.24.2
tk==0.1.0
Pillow==8.3.2
opencv-python==4.5.3
numpy==1.21.2
pytesseract==0.3.8
matplotlib==3.4.3
